backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-core
fastapi
uvicorn[standard]
mcp
pydantic
sqlalchemy
aiosqlite
python-dotenv
